otel/python_receiver/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
googleapis-common-protos==1.70.0
h11==0.16.0
httptools==0.6.4
idna==3.10
opentelemetry-proto==1.32.1
protobuf==5.29.4
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
